llint: check full notCellMask for is_object and is_cell_with_type
These operations previously checked only the low byte of notCellMask
when determining whether a JavaScript value is a pointer or not. The top
byte of the mask has the pattern that identifies pointers, which was not
being checked. In sum, these bugs caused all integers that did not have
the second bit set to be interpreted as pointers for the affected
operations. They are unrelated to CHERIfication, but were discovered by
CHERI protection failures that resulted running the ECMAScript 6 test
suite. They have already been fixed upstream.